// src/app/core/models/admin.model.ts
export interface AdminExercise {
  id: number;
  category_id: number;
  title: string;
  description: string;
  difficulty: string;
  solution?: string | null;
}